+++ changelog.md
@@ -2,7 +2,7 @@ v0.0.7-alpha
 ------------
 * Add Mana-tinged Arrows, an early-game improvement to ammunition
 * Fix failure to clone player mod data upon death
-* Rearrange basic tier research
+* Rearrange research tree
 
 v0.0.6-alpha
 ------------